remove quantization

+++ examples/sequence/include/Decoder_LDPC_BP_flooding_cuda.hxx
@@ -62,7 +62,7 @@ template<typename B, typename R>
 int
 Decoder_LDPC_BP_flooding_cuda<B, R>::_decode_siho_gpu(const B* Y_N, int8_t* CWD, R* V_K, const size_t frame_id)
 {
-    sp_cuda::ldpc_decode(this->ctx,static_cast<const int8_t*>(Y_N),
+    sp_cuda::ldpc_decode(this->ctx,static_cast<const float*>(Y_N),
 				this->K,
 				this->n_ite,
 				0,
